requirements.txt
# Core Flask dependencies
Flask==3.0.0
python-dotenv==1.0.0
gunicorn==21.2.0

# LLM API libraries (optional - install the one you need)
openai>=1.0.0
anthropic>=0.7.0

# Optional auth/security (from original Campus Hub)
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
bcrypt==4.1.1

# Testing
pytest==7.4.3
pytest-flask==1.3.0
